Rename lingering v2 components
