--- src/components/v2v3/V2V3Project/V2V3PayButton/V2PayForm/index.ts
@@ -1,1 +1,0 @@
-export * from './V2PayForm'
