+++ src/components/v2v3/V2V3Project/V2V3PayButton/V2V3ConfirmPayModal/index.ts
@@ -1,0 +1,1 @@
+export { V2V3ConfirmPayModal } from './V2V3ConfirmPayModal'
